chore(rename): rename earn folders
