+++ packages/blockchain-wallet-v4-frontend/src/data/components/interest/types.ts
@@ -91,7 +91,7 @@ export type ErrorStringType = { error: string }
 
 export type InterestLimits = { coin: CoinType; currency: FiatType }
 
-export type EarnTabTypes = 'All' | 'Staking' | 'Rewards'
+export type EarnTabsType = 'All' | 'Staking' | 'Rewards'
 
 export type EarnInstrumentsType = Array<{
   coin: CoinType
@@ -124,7 +124,7 @@ export interface InterestState {
   earnDepositLimits: EarnMinMaxType
   earnEDDStatus: RemoteDataType<string, EarnEDDStatus>
   earnEDDWithdrawLimits: RemoteDataType<string, WithdrawLimits>
-  earnTab: EarnTabTypes
+  earnTab: EarnTabsType
   instruments: RemoteDataType<string, EarnInstrumentsType>
   interestEligible: RemoteDataType<string, EarnEligibleType>
   interestLimits: RemoteDataType<string, InterestLimitsType>
