+++ packages/blockchain-wallet-v4-frontend/src/scenes/Earn/Table/SortableTable/SortableTable.types.ts
