+++ packages/blockchain-wallet-v4-frontend/src/scenes/Earn/Table/Table.model.tsx
